Folding actionnode into vt/topo.
