+++ go/vt/wrangler/shard.go
@@ -7,7 +7,6 @@ package wrangler
 import (
 	"fmt"
 
-	"github.com/youtube/vitess/go/vt/tabletmanager/actionnode"
 	"github.com/youtube/vitess/go/vt/topo"
 	"github.com/youtube/vitess/go/vt/topo/topoproto"
 	"golang.org/x/net/context"
@@ -63,8 +62,8 @@ func (wr *Wrangler) updateShardCellsAndMaster(ctx context.Context, si *topo.Shar
 // This is an emergency manual operation.
 func (wr *Wrangler) SetShardServedTypes(ctx context.Context, keyspace, shard string, cells []string, servedType topodatapb.TabletType, remove bool) (err error) {
 	// lock the keyspace to not conflict with resharding operations
-	actionNode := actionnode.SetShardServedTypes(cells, servedType)
-	ctx, unlock, lockErr := actionNode.LockKeyspace(ctx, wr.ts, keyspace)
+	lock := topo.SetShardServedTypesLock(cells, servedType)
+	ctx, unlock, lockErr := lock.LockKeyspace(ctx, wr.ts, keyspace)
 	if lockErr != nil {
 		return lockErr
 	}
@@ -93,8 +92,8 @@ func (wr *Wrangler) SetShardTabletControl(ctx context.Context, keyspace, shard s
 		return fmt.Errorf("SetShardTabletControl cannot have both DisableQueryService set and tables set")
 	}
 
-	actionNode := actionnode.UpdateShard()
-	ctx, unlock, lockErr := actionNode.LockKeyspace(ctx, wr.ts, keyspace)
+	lock := topo.UpdateShardLock()
+	ctx, unlock, lockErr := lock.LockKeyspace(ctx, wr.ts, keyspace)
 	if lockErr != nil {
 		return lockErr
 	}
@@ -239,8 +238,8 @@ func (wr *Wrangler) RemoveShardCell(ctx context.Context, keyspace, shard, cell s
 
 // SourceShardDelete will delete a SourceShard inside a shard, by index.
 func (wr *Wrangler) SourceShardDelete(ctx context.Context, keyspace, shard string, uid uint32) (err error) {
-	actionNode := actionnode.UpdateShard()
-	ctx, unlock, lockErr := actionNode.LockKeyspace(ctx, wr.ts, keyspace)
+	lock := topo.UpdateShardLock()
+	ctx, unlock, lockErr := lock.LockKeyspace(ctx, wr.ts, keyspace)
 	if lockErr != nil {
 		return lockErr
 	}
@@ -269,8 +268,8 @@ func (wr *Wrangler) SourceShardDelete(ctx context.Context, keyspace, shard strin
 
 // SourceShardAdd will add a new SourceShard inside a shard
 func (wr *Wrangler) SourceShardAdd(ctx context.Context, keyspace, shard string, uid uint32, skeyspace, sshard string, keyRange *topodatapb.KeyRange, tables []string) (err error) {
-	actionNode := actionnode.UpdateShard()
-	ctx, unlock, lockErr := actionNode.LockKeyspace(ctx, wr.ts, keyspace)
+	lock := topo.UpdateShardLock()
+	ctx, unlock, lockErr := lock.LockKeyspace(ctx, wr.ts, keyspace)
 	if lockErr != nil {
 		return lockErr
 	}
